catalog/requirements.txt
bson==0.5.8
grpcio==1.19.0
protobuf==3.7.1
pymongo==3.7.2
python-dateutil==2.8.0
python-dotenv==0.10.1
six==1.12.0
